Refactor: 라우터 테스트 앱 분리 #192
