+++ test/app/router/app_router_test.dart
@@ -1,156 +1,18 @@
-import 'package:commonplant_frontend/app/common_plant_app.dart';
-import 'package:commonplant_frontend/app/router/app_route_spec.dart';
 import 'package:commonplant_frontend/app/router/app_router.dart';
-import 'package:commonplant_frontend/app/router/app_routes.dart';
-import 'package:commonplant_frontend/app/router/route_parameter_error_page.dart';
-import 'package:commonplant_frontend/app/router/route_parameters.dart';
 import 'package:commonplant_frontend/app/router/route_paths.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import '../../helpers/test_app.dart';
+
 void main() {
-  test('Figma 기준 route-level screen 18개를 등록한다', () {
-    expect(appRouteSpecs, hasLength(18));
-    expect(appRouteSpecs.map((route) => route.name).toSet(), hasLength(18));
-    expect(appRouteSpecs.map((route) => route.path).toSet(), hasLength(18));
-  });
-
-  test('필수 path parameter는 null, 빈 문자열, 공백 문자열을 허용하지 않는다', () {
-    expect(
-      requiredPathParameter(const {'placeId': 'place-1'}, 'placeId'),
-      'place-1',
-    );
-    expect(requiredPathParameter(const {}, 'placeId'), isNull);
-    expect(requiredPathParameter(const {'placeId': ''}, 'placeId'), isNull);
-    expect(requiredPathParameter(const {'placeId': '   '}, 'placeId'), isNull);
-  });
-
-  test('선택 query parameter는 값이 있을 때 trim한 문자열을 반환한다', () {
-    expect(optionalQueryParameter(const {}, 'placeId'), isNull);
-    expect(optionalQueryParameter(const {'placeId': ''}, 'placeId'), isNull);
-    expect(optionalQueryParameter(const {'placeId': '   '}, 'placeId'), isNull);
-    expect(
-      optionalQueryParameter(const {'placeId': ' place-1 '}, 'placeId'),
-      'place-1',
-    );
-  });
-
-  testWidgets('route parameter 오류 화면이 누락된 parameter와 route 정보를 표시한다', (
-    WidgetTester tester,
-  ) async {
-    const route = AppRouteSpec(
-      name: AppRouteNames.placeDetail,
-      path: AppRoutePaths.placeDetail,
-      title: '장소 상세',
-      figmaFrames: ['#2-3 My place 리더화면'],
-    );
-
-    await tester.pumpWidget(
-      const MaterialApp(
-        home: RouteParameterErrorPage(route: route, parameterName: 'placeId'),
-      ),
-    );
-
-    expect(find.text('라우트 정보를 확인할 수 없어요'), findsOneWidget);
-    expect(find.text('필수 경로 값이 누락되었습니다: placeId'), findsOneWidget);
-    expect(find.text('routeName: placeDetail'), findsOneWidget);
-    expect(find.text('path: /places/:placeId'), findsOneWidget);
-  });
-
-  testWidgets('Figma 기준 route-level screen이 모두 실제 화면으로 진입된다', (
-    WidgetTester tester,
-  ) async {
-    final locations = <String>[
-      AppRoutePaths.home,
-      AppRoutePaths.onboarding,
-      AppRoutePaths.login,
-      AppRoutePaths.profileSetup,
-      AppRoutePaths.terms,
-      AppRoutePaths.placeInvitations,
-      AppRoutePaths.placeCreate,
-      AppRoutePaths.addressSearch,
-      AppRoutePaths.placeFriendAdd,
-      AppRoutePaths.placeEditLocation('place-1'),
-      AppRoutePaths.friendManagementLocation('place-1'),
-      AppRoutePaths.placeDetailLocation('place-1'),
-      AppRoutePaths.plantSearch,
-      AppRoutePaths.plantCreateDetails,
-      AppRoutePaths.plantEditLocation('plant-1'),
-      AppRoutePaths.memoWriteLocation('plant-1'),
-      AppRoutePaths.memoListLocation('plant-1'),
-      AppRoutePaths.plantDetailLocation('plant-1'),
-    ];
-
-    for (final location in locations) {
-      final router = createAppRouter(initialLocation: location);
-
-      await tester.pumpWidget(
-        ProviderScope(
-          overrides: [appRouterProvider.overrideWithValue(router)],
-          child: const CommonPlantApp(),
-        ),
-      );
-      await tester.pumpAndSettle();
-
-      expect(find.text('라우트 준비 중'), findsNothing);
-
-      await tester.pumpWidget(const SizedBox.shrink());
-      router.dispose();
-    }
-  });
-
-  testWidgets('place detail route가 실제 화면을 표시한다', (WidgetTester tester) async {
-    final router = createAppRouter(
-      initialLocation: AppRoutePaths.placeDetailLocation('place-1'),
-    );
-    addTearDown(router.dispose);
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
-    await tester.pumpAndSettle();
-
-    expect(find.text('My place'), findsOneWidget);
-    expect(find.text('스윗 홈_거실'), findsOneWidget);
-    expect(find.text('서울시 노원구 광운로 20'), findsOneWidget);
-    expect(find.text('라우트 준비 중'), findsNothing);
-  });
-
-  testWidgets('memo route가 실제 작성 화면을 표시한다', (WidgetTester tester) async {
-    final router = createAppRouter(
-      initialLocation: AppRoutePaths.memoWriteLocation('plant-1'),
-    );
-    addTearDown(router.dispose);
-
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
-    await tester.pumpAndSettle();
-
-    expect(find.text('메모 작성'), findsOneWidget);
-    expect(find.text('메모 내용을 입력해 주세요'), findsOneWidget);
-    expect(find.text('라우트 준비 중'), findsNothing);
-  });
-
   testWidgets('onboarding 시작 플로우가 profile, terms, home으로 이어진다', (
     WidgetTester tester,
   ) async {
     final router = createAppRouter(initialLocation: AppRoutePaths.onboarding);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.ensureVisible(find.text('시작하기'));
@@ -188,12 +50,7 @@ void main() {
     final router = createAppRouter(initialLocation: AppRoutePaths.profileSetup);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.text('보기'));
@@ -214,12 +71,7 @@ void main() {
     final router = createAppRouter();
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.text('장소 추가하기'));
@@ -235,12 +87,7 @@ void main() {
     final router = createAppRouter(initialLocation: AppRoutePaths.placeCreate);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.enterText(find.byType(TextField), '옥상 정원');
@@ -274,12 +121,7 @@ void main() {
     final router = createAppRouter(initialLocation: AppRoutePaths.placeCreate);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.enterText(find.byType(TextField), '옥상 정원');
@@ -316,12 +158,7 @@ void main() {
     final router = createAppRouter(initialLocation: AppRoutePaths.placeCreate);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.text('주소'));
@@ -357,12 +194,7 @@ void main() {
     );
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.bySemanticsLabel('장소 상세 메뉴'));
@@ -394,12 +226,7 @@ void main() {
     final router = createAppRouter(initialLocation: AppRoutePaths.plantSearch);
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.enterText(find.byType(TextField), '몬스테라');
@@ -420,12 +247,7 @@ void main() {
     );
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.byTooltip('식물 상세 메뉴'));
@@ -465,12 +287,7 @@ void main() {
     );
     addTearDown(router.dispose);
 
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: [appRouterProvider.overrideWithValue(router)],
-        child: const CommonPlantApp(),
-      ),
-    );
+    await tester.pumpWidget(buildCommonPlantRouterTestApp(router));
     await tester.pumpAndSettle();
 
     await tester.tap(find.text('작성하기'));
