+++ test/features/place/presentation/pages/place_detail_page_test.dart
@@ -11,10 +11,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
 
+import '../../../../helpers/test_app.dart';
+
 void main() {
   testWidgets('리더는 FAB에서 장소 수정 액션을 확인할 수 있다', (tester) async {
     await tester.pumpWidget(
-      _buildPlaceDetailApp(
+      buildPageTestApp(
         const PlaceDetailPage(placeId: 'place-1', role: PlaceDetailRole.leader),
       ),
     );
@@ -30,7 +32,7 @@ void main() {
 
   testWidgets('팀원은 FAB에서 장소 수정 액션을 보지 않는다', (tester) async {
     await tester.pumpWidget(
-      _buildPlaceDetailApp(
+      buildPageTestApp(
         const PlaceDetailPage(placeId: 'place-1', role: PlaceDetailRole.member),
       ),
     );
@@ -46,7 +48,7 @@ void main() {
 
   testWidgets('장소 나가기는 확인 dialog를 표시한다', (tester) async {
     await tester.pumpWidget(
-      _buildPlaceDetailApp(const PlaceDetailPage(placeId: 'place-1')),
+      buildPageTestApp(const PlaceDetailPage(placeId: 'place-1')),
     );
     await tester.pumpAndSettle();
 
@@ -148,10 +150,6 @@ void main() {
     expect(find.text('홈'), findsOneWidget);
     expect(tester.takeException(), isNull);
   });
-}
-
-Widget _buildPlaceDetailApp(Widget home) {
-  return ProviderScope(child: MaterialApp(home: home));
 }
 
 Widget _remotePlaceDetailApp(PlaceRepository repository) {
